Only generate JWT secret if not already present via environment variable

+++ docker-entrypoint.sh
@@ -85,8 +85,15 @@ then
     error_with_msg "MariaDB did not stop"
 fi
 
-echo "- Generating JWT key"
-dd if=/dev/urandom bs=32 count=1 >/etc/slurm/jwt/jwt_hs256.key
+if [ "$JWT_SECRET" ]
+then
+    echo "- JWT secret variable found, writing..."
+    echo "$JWT_SECRET" > /etc/slurm/jwt/jwt_hs256.key
+else
+    echo "- No JWT secret variable found, generating JWT key"
+    dd if=/dev/urandom bs=32 count=1 >/etc/slurm/jwt/jwt_hs256.key
+fi
+
 chown slurm:slurm /etc/slurm/jwt/jwt_hs256.key
 chmod 0600 /etc/slurm/jwt/jwt_hs256.key
